Expose an exception-raising entry point (#32)
`protovalidate.validate` is the main entry point for this package, but
it's not particularly Pythonic: it returns a collection of violations
rather than raising an exception. This commit splits the package entry
point into `validate` and `collect_violations`, shunting most
performance and protobuf-specific concerns to `collect_violations`.

`validate` is now more similar to other Python validators: it takes a
message and raises a `ValidationError` if the message is invalid. The
exception exposes the raw protobuf `Violations` object (in case it needs
to be sent over the network as an error detail), but it also exposes an
`errors()` method that returns a simple Python list of violations. This
matches `pydantic`, the most popular validation library in Python, as
closely as we can without controlling the protobuf-generated
initializers.

`collect_violations` is now the lower-level entry point. It can
optionally append violations to a user-provided `Violations`, and it
lets the user choose whether and how to raise an exception.

To my eye, this makes simple use of this package feel more like normal
Python while still permitting more complex use.