Explicitly support only Python 3.11+
There are a bunch of dev dependency problems trying to run on older
Python versions in CI. In practice, we only support Python 3.11+
today, so we should make that explicit.

In the future, it'd be nice to sort out the CI issues so that we can
support older Python versions too.