Fix behavior of required + ignore_empty constraints (#97)
Brings the python library in conformance with the revised spec around
`required` and `ignore_empty` constraints. See
bufbuild/protovalidate#115

The two `# type: ignore` comments are required as typeshed does not
currently recognize the `has_presence` property on FieldDescriptor
class. I may open an upstream patch to add it in (and audit the rest of
the protobuf type definitions).