Increase conformance test suite timeout
Python is not particularly fast. The new test suites seem to run close
to the timeout on GitHub actions, especially on older, slower Python
versions, so let's increase the timeout a bit.

While the Python protovalidate implementation could be faster, this is
probably also reflective of GitHub Actions being somewhat slow, too.