Implement workaround for field presence issues
Right now, it seems like it is not possible to implement field presence
100% accurately for protobuf types when using cel python, as the result
of the `has` macro is tied to whether or not the expression in it can be
evaluated. This fails in the protobuf case because `has` should be false
in some cases even when a default value will be returned.

This is a somewhat ugly workaround that patches the interpreter so we
can track when we are inside of a `has` macro call and cause there to be
an evaluation error if the field is unset.