Update for editions (#174)
Updates protovalidate, adds support for new ignore semantics. The PR is
huge but it's mostly gencode; the actual CEL code wasn't too
substantially altered.

~~Unfortunately I could not get the tests to pass 100%. There is a minor
wrinkle where, I _believe_, the behavior of `has(x.y)` and `x.y` needs
to differ (specifically with proto3 optional) but it can not due to the
way the interpreter works. It _seems_ like this only impacts some rather
small edge cases so I guess we might be able to live with it in the
interim. I've just added said tests to the nonconforming list.~~
**Update:** The issues still exist upstream (assuming the issue is valid
and not a mistake on my part) but we are now working around it, and
tests now pass 100% (sans the existing failures.)

All other tests pass and I _believe_ my fixes are correct. The e-mail
check was corrected to not support bracketed emails as was done with the
Go version.