Enhance `is_uri` and `is_uri_reference` validation (#271)
This ports the validation logic from `protovalidate-go` for validating
URIs and URI references. As a result, all conformance tests for 0.10.3
related to these validations are now passing.

This uses a single underscore for protected class members and a
double-underscore for private class methods to use name-mangling. There
are varying recommendations on how to structure these (for example,
Effective Python recommends public members, some recommendations suggest
a single-underscore for class methods, etc.).

The current pattern in protovalidate-python seems to use a single
underscore for module-level functions, which we could certainly adopt
for classes. I'm open to whatever is the established pattern that we
should adopt.