Move unit tests to local protos (#288)
The unit tests in this repo use Protobuf definitions from the
conformance suite, which works but it makes debugging issues a bit
difficult. It is extremely helpful to have local protos in the repo that
can be changed for tests to debug issues.

This follows the example in protovalidate-go by creating a
`validations.proto` and copies the definitions currently being used from
conformance into this proto.