Add conformance testing to string.format implementation (#307)
Fixes #257 

This adds conformance testing to our implementation of `string.format`.
cel-python does not currently provide a `string.format` function so we
implement our own and add it to the protovalidate-python CEL runtime.

While formatting is not (yet) a part of the cel-spec, they do provide
conformance test data for validating behavior. So, to ensure we are
conformant, we use this test data and run it against our implementation.

See the test data
[here](https://github.com/google/cel-spec/blob/master/tests/simple/testdata/string_ext.textproto).
Note that this file contains various sections for testing. However, we
are only concerned with the `format` and `format_errors` sections. In
addition, since we are only testing for formatting, we do not handle
every `oneof` case or variation in the `SimpleTest` definition -- only
expected results or evaluation errors. If others are added down the road
in future versions of the spec, we can adjust these tests.