Implement matches override (#319)
This implements an override of the `matches` function in celpy, which
allows us to perform quasi-re2 operations. Since google-re2 [does not
yet support Python 3.13](https://github.com/google/re2/issues/516), we
cannot use it in protovalidate-python.

So, as a good compromise until then, we are simulating re2 behavior much
like protovalidate-es does by failing on syntax not allowed in re2.

Note that protovalidate-es also parses regex flags at the beginning of
regex patterns and applies them at the end. This is because ECMAScript
regex does not support flags at the beginning. However, Python's re
package _does_ support this. So really all this does for now is to
simply fail on invalid re2 syntax.

A future PR will add the ability for users to specify their own re2
engine via a config if they would like to override this behavior (which
would not be a breaking change).

---------

Signed-off-by: Sri Krishna <skrishna@buf.build>
Co-authored-by: Sri Krishna <93153132+srikrsna-buf@users.noreply.github.com>