Use uv_build as build backend
Requires https://github.com/astral-sh/uv/releases/tag/0.7.19.

However, doesn't quite work yet with our release process as we have a
dynamic version, and uv_build does not support getting the version from
git yet. To submit, we'll probably need to revert this, but interesting
to give it a shot.

Ref: https://github.com/astral-sh/uv/issues/14037