Convert from pipenv to uv (#335)
[uv][1] is a better, faster version of pipenv for managing python
projects.

The expectation here is similar to `pipenv`, in that developers will
install `uv` locally; I haven't pinned a particular version as we aren't
doing anything complicated (yet). Can pin a version later if necessary.

The only thing left that we may want to do here is to use `uv` as our
[build backend](https://docs.astral.sh/uv/concepts/build-backend/),
which would eliminate our use of `hatch`, but uv doesn't have [dynamic
version resolution from
vcs](https://github.com/astral-sh/uv/issues/14137) right now (which is
how we currently do our releases), so holding off on that for the time
being.

[1]: https://docs.astral.sh/uv/