Drop custom matcher & config
We implemented this so that we could avoid a hard dependency on
google/re2 due to the lack of wheels being published for Python
3.13. However, [3.13 wheels are now published][1], and our downstream
cel-python dep, as of v0.4, is requiring google/re2 as a dependency. It
seems like we ought to remove our matches workaround, as well as the
configuration for supplying a user-defined matching func, as we
generally want all users to have the same behavior for matching. (We can
keep in the config value if we really think it's necessary.)

[1]: https://github.com/google/re2/issues/516