Add re2 as optional dependency (#346)
This adds `re2` as an optional dependency to protovalidate-python. With
this configured, users can choose to optionally install re2 if they
explicit re2 functionality.

For example

`pip install protovalidate[re2]`

Note that while CEL requires re2 as part of the spec, our current
version of cel-python (v0.2.0) doesn't. We are unable to upgrade
cel-python beyond v0.2.0 at this point due to some outstanding issues,
so since re2 now supports Python 3.13, we're adding it as an optional
dep directly to protovalidate-python rather than the re2 workaround we
had previously implemented.