Remove collect_violations(..., into=) argument (#365)
At risk of Chesterton's fencing myself, but I don't think this argument
is needed or ever used (other than in tests). I can't think of a reason
why someone would want to shove multiple validations into the same
validation, instead of simply calling `collect_violations` multiple
times and combining the lists afterwards.

I did some `git blame` spelunking and it looks like this originally
turned up in #32, but didn't see much reasoning for it in that PR. The
PR mentioned both Pydantic and making things closer to
protovalidate-go's API, but neither seem to have a similar style of
reusing a returned error/exception like this.

Stacked on #364 as these both munge with the exported API & docstrings;
will tag a `v0.15.0` release once all of this is merged.