Reapply "Pin local Python version to lowest supported (#351)" (#375)
This reverts commit fb158759e4d5364b1669df0187ea0c4005104bde.

When working through #384 without a pin set, running `make` would
default to Python 3.13 locally. (Not sure if this is a `uv` version
thing, or what.)

I think it's best to keep this pinned in Python libraries to the lowest
supported Python version to avoid confusion, especially since we don't
specify a `uv` version to use (for now).